+++ app/domain/labor-market.ts
